StreamLit Dashboard/requirements.txt
binance==0.3
chart_studio==1.1.0
coinmetrics==0.2.5
iso8601==1.0.2
matplotlib==3.3.4
mpl_finance==0.10.1
numpy==1.20.1
pandas==1.3.0
Pillow==8.4.0
plotly==5.3.1
python_dateutil==2.8.2
pyti==0.2.2
regex==2021.4.4
requests==2.25.1
san==1.2.0
sanpy==0.8.11
scipy==1.6.2
seaborn==0.11.1
streamlit==1.1.0
streamlit_authenticator==0.0.5
tqdm==4.59.0
